pyproject.toml
[project]
name = "ecommerce-visual-pro"
version = "0.1.0"
description = "AI-powered e-commerce product photo optimization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "pillow>=10.2.0",
    "python-multipart>=0.0.9",
    "google-cloud-storage>=2.14.0",
    "torch>=2.2.0",
    "transformers",
    "torchvision>=0.17.0",
    "scikit-image>=0.22.0",
    "numpy>=1.26.0",
    "python-jose[cryptography]>=3.5.0",
    "diffusers>=0.31.0",
    "safetensors>=0.4.0",
    "accelerate>=0.27.0",
    "flower>=2.0.1",
    "huggingface-hub[hf-xet]>=0.36.2",
    "gguf>=0.6.0",
    "sentencepiece>=0.2.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disable_error_code = ["untyped-decorator", "no-untyped-call"]
